requirements.txt
beautifulsoup4
pandas
numpy
matplotlib
seaborn
geopandas
folium
python-dotenv
google-auth
google-api-python-client
psycopg2
pyarrow
psycopg2-binary
datetime 
timedelta
os
ast
pyarrow





